Update command does not exist.
Having an update command enables users to update specific
details of their task without having to delete and add
their task again which is troublesome.

Lets add a new UpdateCommand that inherits Command.
When executed, this command will update the relevant
task in the task list.

This allows users to update existing tasks by
specifying which task to update and the detail to
update, such as the description or date.